@@ -39,28 +39,34 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
   ionicons: ^0.2.2
-  loading_overlay: ^0.3.0
+  loading_overlay: ^0.4.2
   provider: ^6.1.2
   google_fonts: ^6.2.1
   supabase_flutter: ^2.6.0
-  bloc: ^8.1.4
-  flutter_bloc: ^8.1.6
+  bloc: ^9.0.0
+  flutter_bloc: ^9.1.1
   meta: ^1.12.0
   equatable: ^2.0.5
   uuid: ^4.4.2
-  formz: ^0.7.0
+  formz: ^0.8.0
   flow_builder: ^0.1.0
   modal_bottom_sheet: ^3.0.0
   form_builder_validators: ^11.0.0
-  font_awesome_flutter: ^10.7.0
-  flutter_form_builder: ^9.5.0
-  intl: ^0.19.0
+  font_awesome_flutter: ^10.9.0
+  flutter_form_builder: ^10.0.1
+  intl: ^0.20.2
   expandable_text: ^2.3.0
   searchfield: ^1.1.8
-  flutter_map: ^6.1.0
   latlong2: ^0.9.0
   geodesy: ^0.10.2
-  geolocator: ^10.0.0
+  geolocator: ^14.0.1
+  firebase_messaging: ^16.0.1
+  flutter_map: ^8.1.1
+  #flutter_map_location_picker: ^0.0.4
+  geocoding: ^4.0.0
+  logging: ^1.3.0
+  http: ^1.5.0
+  timeago: ^3.7.1
 
 dev_dependencies:
   flutter_test:
@@ -73,7 +79,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^4.0.0
+  flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
