@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart' show SystemMouseCursor;
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:ionicons/ionicons.dart';
@@ -8,15 +9,17 @@ import 'package:support_sphere/logic/cubit/home_state.dart';
 import 'package:support_sphere/logic/bloc/auth/authentication_bloc.dart';
 import 'package:support_sphere/presentation/components/home/home_header.dart';
 import 'package:support_sphere/presentation/components/home/home_map.dart';
+import 'package:geodesy/geodesy.dart';
+
 
 class HomeBody extends StatefulWidget {
   const HomeBody({super.key});
 
   @override
-  State<HomeBody> createState() => _HomeBodyState();
+  State<HomeBody> createState() => HomeBodyState();
 }
 
-class _HomeBodyState extends State<HomeBody> {
+class HomeBodyState extends State<HomeBody> {
   late final MapController _mapController;
   bool _isMapReady = false;
 
@@ -28,7 +31,7 @@ class _HomeBodyState extends State<HomeBody> {
 
   @override
   Widget build(BuildContext context) {
-    final AuthUser authUser = context.select(
+    final MyAuthUser authUser = context.select(
       (AuthenticationBloc bloc) => bloc.state.user,
     );
 
@@ -38,6 +41,9 @@ class _HomeBodyState extends State<HomeBody> {
         listener: (context, state) {
           if (state.status == HomeStatus.success) {
             _recenterMap(state);
+          }
+          else if (state.status == HomeStatus.editMeetingPlace) {
+            _editMode(state);
           }
         },
         builder: (context, state) {
@@ -52,22 +58,41 @@ class _HomeBodyState extends State<HomeBody> {
                   if (state.cluster != null)
                     HomeHeader(cluster: state.cluster!),
                   Expanded(
-                    child: HomeMap(
-                      mapController: _mapController,
-                      userLocation: state.userLocation,
-                      initMapCentroid: state.initMapCentroid,
-                      initZoomLevel: state.initZoomLevel,
-                      captainMarkers: state.captainMarkers,
-                      onMapReady: () {
-                        setState(() => _isMapReady = true);
-                      },
+                    child: MouseRegion(
+                      cursor: _cursorFor(state.status),
+                      child: HomeMap(
+                        mapController: _mapController,
+                        state: state,
+                        cubit: context.read<HomeCubit>(),
+                        onMapReady: () {
+                          setState(() => _isMapReady = true);
+                        },
+                      ),
                     ),
                   ),
                 ],
               ),
               Positioned(
-                right: 16,
+                // start edit-mode button
+                left: 16,
                 bottom: 16,
+                child: FloatingActionButton(
+                  onPressed: () {
+                    final cubit = context.read<HomeCubit>();
+                    cubit.editMeetingPlace();
+                  },
+
+                  backgroundColor: Colors.white,
+                  elevation: 2,
+                  child: const Icon(
+                    Ionicons.flag,
+                    color: Colors.black,
+                  ),
+                ),
+              ), // end edit-mode button
+              Positioned(
+                right: 16,
+                bottom: 86,
                 child: FloatingActionButton(
                   onPressed: () async {
                     final cubit = context.read<HomeCubit>();
@@ -84,6 +109,25 @@ class _HomeBodyState extends State<HomeBody> {
                   ),
                 ),
               ),
+              //
+              Positioned(
+                right: 16,
+                bottom: 16,
+                child: FloatingActionButton(
+                  onPressed: () async {
+                    final cubit = context.read<HomeCubit>();
+                    // could flip icon! custom icon? mouse pointer?
+                    // assume toggle on/off
+                    await cubit.showAllClusters(state.status != HomeStatus.allClusters);
+                  },
+                  backgroundColor: Colors.white,
+                  elevation: 2,
+                  child: const Icon(
+                    Ionicons.square_outline,
+                    color: Colors.black,
+                  ),
+                ),
+              ),
             ],
           );
         },
@@ -94,6 +138,47 @@ class _HomeBodyState extends State<HomeBody> {
   void _recenterMap(HomeState state) {
     if (!_isMapReady) return;
 
-    _mapController.move(state.userLocation ?? state.initMapCentroid, state.initZoomLevel);
+    _mapController.move(
+        state.userLocation ?? state.initMapCentroid, state.initZoomLevel);
+  }
+
+  void _editMode(HomeState state) {
+    // change icon
+    if (state.cluster != null && state.cluster!.geom != null ) {
+      LatLngBounds? bounds = LatLngBounds.fromPoints(state.cluster!.geom!);
+      _mapController.fitCamera(CameraFit.bounds(bounds: bounds));
+    }
+  }
+
+  LatLng _initMapCentroid(HomeState state) {
+    // first, check user location
+    if (state.userLocation != null) {
+      return state.userLocation!;
+    }
+    if (state.cluster != null) {
+      LatLng? centroid = state.cluster!.centroid();
+      if (centroid != null) {
+        return centroid;
+      }
+    }
+    return LatLng(47.661322762238285, -122.2772993912835);
+  }
+}
+
+// Noting to self, and for posteriety:
+// These state-to-visual mappings could be stored
+// in a DB or simple lookup table.
+SystemMouseCursor _cursorFor(HomeStatus status) {
+  switch (status) {
+    case HomeStatus.initial:
+    case HomeStatus.loading:
+      return SystemMouseCursors.wait;
+    case HomeStatus.editMeetingPlace:
+      return SystemMouseCursors.grabbing;
+    case HomeStatus.success:
+    case HomeStatus.allClusters:
+      return SystemMouseCursors.basic;
+    case HomeStatus.failure:
+      return SystemMouseCursors.forbidden;
   }
 }
