@@ -2,7 +2,7 @@ import logging
 from importlib.metadata import version
 
 logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s %(thread)d %(levelname)s %(module)s.%(funcName)s(): %(message)s"
+    level=logging.DEBUG, format="%(asctime)s %(thread)d %(levelname)s %(module)s.%(funcName)s(): %(message)s"
 )
 logger = logging.getLogger(__name__)
 
