@@ -2,20 +2,26 @@ import csv
 import datetime
 import uuid
 import time
+from support_sphere.models.public.point_of_interest import PointOfInterest, PointOfInterestType
+from support_sphere.models.public.message import Message
 import typer
 import json
+from shapely.wkt import loads
+
 
 from pathlib import Path
 
 from support_sphere.models.public import (UserProfile, People, Cluster, PeopleGroup, Household,
                                           RolePermission, UserRole, UserCaptainCluster, SignupCode,
-                                          ResourceType, ResourceCV, Resource, Checklist, ChecklistStep, 
+                                          ResourceType, ResourceCV, Resource, Checklist, ChecklistStep,
                                           ChecklistStepsOrder, Frequency)
 from support_sphere.models.auth import User
 from support_sphere.repositories.auth import UserRepository
 from support_sphere.repositories.base_repository import BaseRepository
 from support_sphere.repositories.public import UserProfileRepository, UserRoleRepository, PeopleRepository
 from support_sphere.repositories import supabase_client
+from geoalchemy2.shape import from_shape
+from shapely.geometry import shape
 
 from support_sphere.models.enums import AppRoles, AppPermissions, OperationalStatus
 
@@ -23,7 +29,7 @@ import logging
 
 DATA_DIRECTORY = Path(__file__).parent / 'resources' / 'data'
 
-logger = logging.getLogger(__name__)
+log = logging.getLogger(__name__)
 
 db_init_app = typer.Typer()
 
@@ -90,17 +96,19 @@ def populate_user_details():
         for row in csv_reader:
             user_profile = None
             if bool(eval(row['has_profile'])):
-                # Create a auth.user with encrypted_password (ONLY FOR LOCAL TESTING)
-                supabase_client.auth.sign_up({"email": row['email'], "password": row['username']})
-                supabase_client.auth.sign_out()
                 user: User = UserRepository.find_by_email(row['email'])
+                if not user:
+                    # Create a auth.user with encrypted_password (ONLY FOR LOCAL TESTING)
+                    supabase_client.auth.sign_up({"email": row['email'], "password": row['username']})
+                    supabase_client.auth.sign_out()
+                    user: User = UserRepository.find_by_email(row['email'])
 
-                # Create a user profile
-                profile = UserProfile(user=user)
-                user_profile = UserProfileRepository.add(profile)
+                    # Create a user profile
+                    profile = UserProfile(user=user)
+                    user_profile = UserProfileRepository.add(profile)
 
-                user_role = UserRole(user_profile=user_profile, role=AppRoles.USER)
-                BaseRepository.add(user_role)
+                    user_role = UserRole(user_profile=user_profile, role=AppRoles.USER)
+                    BaseRepository.add(user_role)
 
             # Create People Entry
             person_detail = People(given_name=row['given_name'], family_name=row['family_name'],
@@ -113,7 +121,8 @@ def populate_user_details():
             # Create a PeopleGroup Entry
             people_group = PeopleGroup(people=person, household=all_households[-1])
             BaseRepository.add(people_group)
-    logger.info("Database Populated Successfully")
+    log.info("Database Populated Successfully")
+
 
 def populate_checklists():
     """
@@ -141,15 +150,28 @@ def populate_checklists():
             step_order = ChecklistStepsOrder(checklist_id=checklist.id, checklist_step_id=step.id, priority=idx)
             BaseRepository.add(step_order)
 
+
 @db_init_app.command(help="Setup a dummy cluster and a household")
 def populate_cluster_and_household_details():
     # Creating entries in 'Cluster' and 'Household' table.
-    cluster = Cluster(name="Cluster1")
-    BaseRepository.add(cluster)
-    all_clusters = BaseRepository.select_all(Cluster)
+    # load a file of Clusters
+    # name: str|None = Field(nullable=True) < !!!
+    # meeting_place: str|None = Field(nullable=True)
+    # notes: str | None = Field(nullable=True)
+    # geom: Geometry|None = Field(sa_type=Geometry(geometry_type="POLYGON"), nullable=True) < !!!
 
-    household = Household(cluster=all_clusters[-1], name="Household1")
-    BaseRepository.add(household)
+    for cluster in load_clusters():
+        try:
+            BaseRepository.add(cluster)
+        except Exception as ex:
+            log.error(f"Error adding {cluster}: {ex}")
+
+    for cluster in BaseRepository.select_all(Cluster):
+        if cluster.name == "c_1":
+            household = Household(cluster=cluster, name="Household1")
+            BaseRepository.add(household)
+            log.info(f"added household {household}")
+            break
 
 
 def generate_signup_codes(household_id: uuid.UUID):
@@ -168,7 +190,7 @@ def generate_signup_codes(household_id: uuid.UUID):
             # Add signup code to the database
             BaseRepository.add(signup_code)
         except Exception as e:
-            logger.error(f"Error: {e}... trying again")
+            log.error(f"Error: {e}... trying again")
             time.sleep(2)
             continue
         break
@@ -185,18 +207,16 @@ def populate_real_cluster_and_household():
         csv_reader = csv.DictReader(file)
 
         cluster_uids = {}
+        for cluster in BaseRepository.select_all(Cluster):
+            cluster_uids[cluster.name] = cluster.id
+
         for row in csv_reader:
             # Get and set cluster
             cluster_name = row["CLUSTER"]
-            if cluster_name not in cluster_uids:
-                cluster = Cluster(name=cluster_name)
-                cluster_id = cluster.id
-                cluster_uids[cluster_name] = cluster.id
-
-                # Add cluster to the database
-                BaseRepository.add(cluster)
-            else:
+            if cluster_name in cluster_uids:
                 cluster_id = cluster_uids[cluster_name]
+            else:
+                log.error(f"Unknown cluster name: {cluster_name}")
 
             # Setup household
             household_address = row['ADDRESS']
@@ -210,11 +230,14 @@ def populate_real_cluster_and_household():
 
 @db_init_app.command(help="Sanity check for sign-up and sign-in via supabase")
 def authenticate_user_signup_signin_signout_via_supabase():
-    # The password is stored in an encrypted format in the auth.users table
-    response_sign_up = supabase_client.auth.sign_up({"email": "zeta@abc.com", "password": "zetazeta"})
-    supabase_client.auth.sign_out()
-    response_sign_in = supabase_client.auth.sign_in_with_password({"email": "zeta@abc.com", "password": "zetazeta"})
-    supabase_client.auth.sign_out()
+    try:
+        # The password is stored in an encrypted format in the auth.users table
+        supabase_client.auth.sign_up({"email": "zeta@abc.com", "password": "zetazeta"})
+        supabase_client.auth.sign_out()
+        supabase_client.auth.sign_in_with_password({"email": "zeta@abc.com", "password": "zetazeta"})
+        supabase_client.auth.sign_out()
+    except Exception as ex:
+        log.error('Error in %s', 'authenticate_user_signup_signin_signout_via_supabase', exc_info=ex)
 
 
 def update_user_permissions_roles_by_cluster():
@@ -246,7 +269,7 @@ def update_user_permissions_roles_by_cluster():
 
 
 def test_app_mode_status_update():
-    response_sign_in = supabase_client.auth.sign_in_with_password(
+    supabase_client.auth.sign_in_with_password(
         {"email": "beth.bodmas@example.com", "password": "bethbodmas"})
 
     user = UserRepository.find_by_email('beth.bodmas@example.com')
@@ -269,7 +292,7 @@ def test_app_mode_status_update():
 
 def test_unauthorized_app_mode_update():
     try:
-        response_sign_in = supabase_client.auth.sign_in_with_password(
+        supabase_client.auth.sign_in_with_password(
             {"email": "adam.abacus@example.com", "password": "adamabacus"})
         user = UserRepository.find_by_email('adam.abacus@example.com')
         supabase_client.table("operational_events").insert({"id": str(uuid.uuid4()),
@@ -277,8 +300,8 @@ def test_unauthorized_app_mode_update():
                                                             "created_at": datetime.datetime.now().isoformat(),
                                                             "status": OperationalStatus.EMERGENCY.name}).execute()
     except Exception as ex:
-        logger.info(ex)
-        logger.info("[CORRECT BEHAVIOUR]: User Denied Access for missing AUTHz.")
+        log.info(ex)
+        log.info("[CORRECT BEHAVIOUR]: User Denied Access for missing AUTHz.")
     finally:
         supabase_client.auth.sign_out()
 
@@ -296,6 +319,89 @@ def setup_user_details():
     update_user_permissions_roles_by_cluster()
 
 
+@db_init_app.command(help="Setup the datebase with points of interest and supporting types")
+def setup_points_of_interest():
+    populate_point_of_interest_types(DATA_DIRECTORY / 'point_types.csv')
+    populate_points_of_interest()
+
+
+def populate_point_of_interest_types(csv_file: str) -> None: # could return a list of items
+    # load types from file
+    with csv_file.open(mode='r', newline='') as file:
+        csv_reader = csv.DictReader(file)
+        for row in csv_reader:
+            name = row['name']
+            icon = row['icon']
+            point_type = PointOfInterestType(name=name, icon=icon)
+            try:
+                BaseRepository.add(point_type)
+            except Exception as ex:
+                log.error(f"Error adding {name}: {ex}")
+
+
+def populate_points_of_interest():
+    file_path = DATA_DIRECTORY / 'point_of_interests.csv'
+    with file_path.open(mode='r', newline='') as file:
+        csv_reader = csv.DictReader(file)
+        for row in csv_reader:
+            geom = from_shape(shape(json.loads(row['geom'])))
+            point = PointOfInterest(
+                name=row['name'],
+                address=row['address'],
+                geom=geom,
+                point_type_name=row['type'])
+            BaseRepository.add(point)
+
+
+def load_clusters(csv_file: Path = DATA_DIRECTORY/'cluster_map.csv') -> list[Cluster]:
+    clusters = []
+    try:
+        with csv_file.open(mode='r', newline='') as file:
+            csv_reader = csv.DictReader(file)
+            for row in csv_reader:
+                geom = from_shape(loads(row['geom']))
+                cluster = Cluster(
+                    name=row['name'],
+                    meeting_place=row['name'] + " meeting place",
+                    meeting_point = None, # FIXME!
+                    notes = row.get("notes", ""),
+                    geom = geom,
+                )
+                clusters.append(cluster)
+        log.info(f"Loaded {len(clusters)} clusters")
+    except Exception as e:
+        log.error(f"Error loading {csv_file}: {e}")
+    return clusters
+
+
+def load_test_messages(csv_file: Path = DATA_DIRECTORY/'messages.csv') -> list[Message]:
+    messages = []
+    try:
+        with csv_file.open(mode='r', newline='') as file:
+            csv_reader = csv.DictReader(file)
+            for row in csv_reader:
+                from_user: User = UserRepository.find_by_email(row['from_id'])
+                if from_user:
+                    row['from_id'] = from_user.id
+                to_user: User = UserRepository.find_by_email(row['to_id'])
+                if to_user:
+                    row['to_id'] = to_user.id
+                msg = Message.fromDict(row)
+                messages.append(msg)
+        log.info(f"Loaded {len(messages)} messages")
+    except Exception as e:
+        log.error(f"Error loading {csv_file}: {e}")
+    return messages
+
+
+def populate_messages():
+    for message in load_test_messages():
+        try:
+            BaseRepository.add(message)
+        except Exception as ex:
+            log.error(f"Error adding {message}: {ex}")
+
+
 @db_init_app.command(help="Sanity check for testing authorization for app mode change")
 def test_app_mode_change():
     test_app_mode_status_update()
@@ -305,9 +411,10 @@ def test_app_mode_change():
 @db_init_app.command(help="Command to setup the database with "
                           "dummy users, roles, permissions, households, clusters, and app mode with sanity check")
 def run_all():
-    logger.info("Starting to populate db with sample entries...")
+    log.info("Starting to populate db with sample entries...")
 
     # Sanity check for user sign-up and sign-in flow via supabase
+    # FIXME
     authenticate_user_signup_signin_signout_via_supabase()
 
     # Set up a dummy cluster and a household
@@ -319,12 +426,20 @@ def run_all():
     # Setup utility resources to be shared during emergency
     setup_utility_resources()
 
+    # Setup points-of-interest and types
+    setup_points_of_interest()
+
     # Sanity check app mode update
+    # FIXME
     test_app_mode_change()
 
     # Populate real data
     populate_real_cluster_and_household()
-    logger.info("Completed Successfully!")
+
+    # Populate test messages
+    populate_messages()
+
+    log.info("Completed Successfully!")
 
 
 if __name__ == '__main__':
